Made basics quries and made them work
